loot refactoring
