+++ Assets/OtherAssets/Coin,Gold and CashBag/Prefabs/CoinPickupParticle.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 400000}
   - component: {fileID: 4451218580315728970}
   - component: {fileID: 1100129129395494421}
+  - component: {fileID: 5315706284199639470}
   m_Layer: 0
   m_Name: CoinPickupParticle
   m_TagString: Untagged
@@ -4870,3 +4871,16 @@ ParticleSystemRenderer:
   m_MeshWeighting2: 1
   m_MeshWeighting3: 1
   m_MaskInteraction: 0
+--- !u!114 &5315706284199639470
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 100000}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a0e8d3e54f044f5ea6a9c50da4ac9263, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _timeDelay: 2
