+++ Assets/Scripts/UserProgress/ProfileProgressService.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using ConfigurationProviders;
 using UniRx;
 using User;
 using WeaponsSettings;
